fix: implement device warning
